requirements.txt
Django>=5.2.0
djangorestframework
faker